Refactor: move nfo maker into correct namespace and move metadataPP patcher into utils
